64bit: BSP plane.dist corruption — retype FPU-temp v19 from double to int
decompilation.  Original 32-bit MSVC code passed the LittleFloat return
value through the x87 stack as a transparent bit-pattern round-trip,
which IDA naturally rendered as 'double'.

On aarch64 GCC reads the type literally and emits scvtf (signed int ->
double) before the store and fcvtzu (double -> unsigned int) after the
load.  fcvtzu *saturates negative doubles to 0*, so any BSP plane with
negative dist (e.g. 0xc4800000 = -1024.0f) silently became 0.

Downstream consequences:
- BSP leaf walk (sub_10003360) diverges from file-truth path at the
  first plane whose dist was negative -> ends in SOLID leaf (cluster=-1).
- AAS_InPVS returns 0 between two valid empty leaves.
- BotEntityVisible reports frac=0 for every target.
- BotFindEnemy never finds candidates; bots never enter combat AI.

The 32-bit Windows MSVC build is unaffected because x87 reinterprets
the bit pattern losslessly.

Fix: retype v19 to int.  LittleFloat is an int -> int identity on LE;
bytes preserved verbatim through the byteswap loop.

Verification (q2dm1, 4 bots, 5 min post-fix):
- PVScall returns real clusters (cl=773, cl=396; previously cl=-1)
- 75 152 'BotEntityVisible VISIBLE iter=0 frac=1.000' events
- AAS_Trace returns valid fractions
- BotFindEnemy candidates > 0 (previously always 0)

Pattern documented for future occurrences in

Source-commit: 41909ce630b35d83d5ad8bcaf71b9ca800eed32e

+++ botlib/botlib.c
@@ -4825,7 +4825,7 @@ int __fastcall sub_10007460(void *a1)
   int v16; // ebx
   char *v17; // esi
   _DWORD *v18; // edi
-  double v19; // st7
+  int v19; // st7 -- IDA decompiled FPU temp as double; LittleFloat is int identity so plain int round-trips bit pattern
   void *v20; // edx
   int v21; // eax
   int v22; // ebp
